+++ Echo/src/utopia/echo/model/request/chat/tool/ToolParameter.scala
@@ -2,7 +2,32 @@ package utopia.echo.model.request.chat.tool
 
 import utopia.flow.collection.immutable.Empty
 import utopia.flow.generic.casting.ValueConversions._
-import utopia.flow.generic.model.immutable.{Constant, Model}
+import utopia.flow.generic.casting.ValueUnwraps._
+import utopia.flow.generic.model.immutable.{Constant, Model, ModelDeclaration}
+import utopia.flow.generic.model.mutable.DataType.StringType
+import utopia.flow.generic.model.template.ModelLike.AnyModel
+
+object ToolParameter
+{
+	// ATTRIBUTES   ------------------
+	
+	private lazy val schema = ModelDeclaration("type" -> StringType)
+	
+	
+	// OTHER    ----------------------
+	
+	/**
+	  * Parses a tool parameter from a model
+	  * @param name Name of this parameter. Expected to be specified as property name.
+	  * @param model Model representing this parameter.
+	  * @param optional Whether this parameter should be optional (default = false)
+	  * @return Parsed parameter. Failure if "type" was missing.
+	  */
+	def parseFrom(name: => String, model: AnyModel, optional: => Boolean = false) =
+		schema.validate(model).map { model =>
+			apply(name, model("type"), model("description"), model("enum").getVector.map { _.getString }, optional)
+		}
+}
 
 /**
   * Describes an input parameter for a tool