+++ Flow/Changes.md
@@ -1,9 +1,18 @@
 # Utopia Flow - List of Changes
 
 ## v2.5.1 (in development)
+### New features
+- Added **UntilExternalFailureIterator**
 ### New methods
+- **ModelLike**
+  - Added `.tryGet(...)`
 - **Seq** (**CollectionExtensions**)
   - Added `.findNextIndexWhere(...)` and `.findPreviousIndexWhere(...)`
+- **Try** (**TryExtensions**)
+  - Added `.mapFailure(...)`
+### Other changes
+- `.allChildrenIterator` in Path (via **FileExtensions**) now returns the paths from top to bottom
+- Minor refactoring in `ModelLike.apply(IterableOnce)`
 
 ## v2.5 - 04.10.2024
 A larger update focusing on pointers. The main changes are: